do not mock any ContainerBuilder methods
The tests are not expected to call any of the methods of the
`ContainerBuilder` class. Therefore, none need to be mocked. This is
needed as PHPUnit would otherwise complain about not being able to mock
the non-existent `ExpressionFunctionProviderInterface` (which is
provided by the ExpressionLanguage component, but is actually an
optional dependency here).